requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
plotly>=5.18.0
pandas>=2.0.0
psutil>=5.9.0
python-dotenv>=1.0.0
reportlab>=4.0.0
requests>=2.31.0
streamlit-extras>=0.3.6
numpy>=1.24.0
